4 files: extract Card.vue to _storyUtils
